requirements.txt
flask==2.0.1
praw==7.4.0
python-dotenv==0.19.0
gunicorn==20.1.0
tiktoken==0.4.0